Inline query-assigned-nodes
