--- packages/reactive-element/src/std-decorators/query-assigned-nodes.ts
@@ -1,71 +1,0 @@
-/**
- * @license
- * Copyright 2017 Google LLC
- * SPDX-License-Identifier: BSD-3-Clause
- */
-
-/*
- * IMPORTANT: For compatibility with tsickle and the Closure JS compiler, all
- * property decorators (but not class decorators) in this file that have
- * an @ExportDecoratedItems annotation must be defined as a regular function,
- * not an arrow function.
- */
-
-import type {ReactiveElement} from '../reactive-element.js';
-import type {Interface} from '../decorators/base.js';
-
-/**
- * Options for the {@linkcode queryAssignedNodes} decorator. Extends the options
- * that can be passed into [HTMLSlotElement.assignedNodes](https://developer.mozilla.org/en-US/docs/Web/API/HTMLSlotElement/assignedNodes).
- */
-export interface QueryAssignedNodesOptions extends AssignedNodesOptions {
-  /**
-   * Name of the slot to query. Leave empty for the default slot.
-   */
-  slot?: string;
-}
-
-/**
- * An accessor decorator that converts a class property into a getter that
- * returns the `assignedNodes` of the given `slot`.
- *
- * Can be passed an optional {@linkcode QueryAssignedNodesOptions} object.
- *
- * Example usage:
- * ```ts
- * class MyElement {
- *   @queryAssignedNodes({slot: 'list', flatten: true})
- *   listItems!: Array<Node>;
- *
- *   render() {
- *     return html`
- *       <slot name="list"></slot>
- *     `;
- *   }
- * }
- * ```
- *
- * Note the type of this property should be annotated as `Array<Node>`. Use the
- * queryAssignedElements decorator to list only elements, and optionally filter
- * the element list using a CSS selector.
- *
- * @category Decorator
- */
-// TODO(justinfagnani): infer a more precise return type when the query is
-// a tagname.
-export const queryAssignedNodes =
-  (options?: QueryAssignedNodesOptions) =>
-  <C extends Interface<ReactiveElement>, V extends ReadonlyArray<Node>>(
-    _target: ClassAccessorDecoratorTarget<C, V>,
-    _context: ClassAccessorDecoratorContext<C, V>
-  ): ClassAccessorDecoratorResult<C, V> => {
-    const {slot} = options ?? {};
-    return {
-      get(this: C): V {
-        const slotSelector = `slot${slot ? `[name=${slot}]` : ':not([name])'}`;
-        const slotEl =
-          this.renderRoot?.querySelector<HTMLSlotElement>(slotSelector);
-        return (slotEl?.assignedNodes(options) ?? []) as unknown as V;
-      },
-    };
-  };
